Added nav obs
I also, disabled some lights and added the obstacle Nav on doors.

The lights though are shining through the walls since it's being blasted. Let me know what you want to do with that. Since we Can't fully Blast the light source if we are I need to duplicate the wall and make it taller to block the light.