HOLY SHIT! lol
Okay i did a lot of changes.

Lights should be fix and i went room from room to make sure the lights doesn't go throuhg the other side of the wall or room.

I also, exapnded the outer wall and made copy of another wall inside the level so it's thicker.

Some walls can't be help so i just extrude the collider.

Fix some gaps on the ground as well. So play the game and let me know if anything else is broken. lol

I tried breaking the game and holy most walls are broken lol.